Modularise history, aliases and load file
