+++ src/replkit/generic_repl.py
@@ -24,6 +24,8 @@ from .repl_commands import (
 )
 from .alias import expand_aliases, handle_alias_command
 from .history_mixin import HistoryMixin
+from .file_loader_mixin import FileLoaderMixin
+from .alias_mixin import AliasMixin
 
 
 class REPLCompleter:
@@ -83,7 +85,7 @@ class REPLCompleter:
         return self.matches[state] if state < len(self.matches) else None
 
 
-class GenericREPL(HistoryMixin):
+class GenericREPL(HistoryMixin, FileLoaderMixin, AliasMixin):
     """A customizable Read-Eval-Print Loop (REPL) for command interpreters.
 
     Attributes:
@@ -137,179 +139,6 @@ class GenericREPL(HistoryMixin):
             UnaliasCommand(),
         ]
 
-    # def add_history_once(self, line: str):
-    #     """Adds a line to readline history if it's not already the last entry."""
-    #     hist_len = readline.get_current_history_length()
-    #     if hist_len == 0 or readline.get_history_item(hist_len) != line:
-    #         readline.add_history(line)
-
-
-    # def handle_alias_command(self, line: str) -> bool:
-    #     """Handles alias creation, listing, and removal.
-
-    #     Args:
-    #         line: A command line starting with .alias or .unalias.
-
-    #     Returns:
-    #         True if the line was an alias command, False otherwise.
-    #     """
-    #     if line.startswith(".alias"):
-    #         parts = line[len(".alias") :].strip()
-    #         if not parts:
-    #             if not self.aliases:
-    #                 print("No aliases defined.")
-    #             else:
-    #                 for name, expr in sorted(self.aliases.items()):
-    #                     print(f"{name} = {expr}")
-    #         else:
-    #             if "=" not in parts:
-    #                 print("Usage: .alias name=expression")
-    #                 return True
-    #             name, expr = map(str.strip, parts.split("=", 1))
-    #             if not name.startswith("@") or not name[1:].isidentifier():
-    #                 print(
-    #                     f"Invalid alias name: '{name}' (must start with '@' and be a valid identifier)"
-    #                 )
-    #                 return True
-    #             if not expr:
-    #                 print("Alias expression cannot be empty.")
-    #                 return True
-    #             try:
-    #                 expr_expanded = expand_aliases(expr, self.aliases)
-    #             except ValueError as e:
-    #                 print(f"Alias error in expression: {e}")
-    #                 return True
-    #             if name in self.aliases:
-    #                 print(
-    #                     f"Alias '{name}' replaced (was: {self.aliases[name]}) → now: {expr_expanded}"
-    #                 )
-    #             else:
-    #                 print(f"Alias added: {name} = {expr_expanded}")
-    #             self.aliases[name] = expr_expanded
-    #         return True
-    #     if line.startswith(".unalias"):
-    #         parts = line[len(".unalias") :].strip()
-    #         if parts in self.aliases:
-    #             del self.aliases[parts]
-    #             print(f"Alias removed: {parts}")
-    #         else:
-    #             print(f"No such alias: {parts}")
-    #         return True
-    #     self.add_history_once(line)
-
-    #     return False
-
-    # def expand_aliases(self, line: str) -> str:
-    #     """Expands all aliases in a given input line.
-
-    #     Raises:
-    #         ValueError: If an alias used in the line is not defined.
-    #     """
-    #     if not self.aliases or "@" not in line:
-    #         return line
-
-    #     def tokenize(expr):
-    #         token_pattern = r"""(\".*?\"|\'.*?\'|\w+|@[a-zA-Z_][\w_]*|[^\s])"""
-    #         return re.findall(token_pattern, expr)
-
-    #     tokens = tokenize(line)
-    #     result = []
-
-    #     for token in tokens:
-    #         if re.fullmatch(r"@[a-zA-Z_]\w*", token):
-    #             if token in self.aliases:
-    #                 result.append(f"({self.aliases[token]})")
-    #             else:
-    #                 raise ValueError(f"Unknown alias: '{token}'")
-    #         else:
-    #             result.append(token)
-
-    #     return " ".join(result)
-
-    # def init_history(self):
-    #     """Initializes the readline history from file."""
-    #     try:
-    #         readline.read_history_file(self.history_file)
-    #         self.logger.info("Loaded history file: %s", self.history_file)
-    #     except FileNotFoundError:
-    #         self.logger.warning("History file not found: %s", self.history_file)
-    #     readline.set_history_length(self.history_length)
-
-    # def save_history(self):
-    #     """Saves the readline history to file."""
-    #     readline.write_history_file(self.history_file)
-    #     self.logger.info("Saved history to: %s", self.history_file)
-
-    # def print_history(self):
-    #     """Prints the current command history to stdout."""
-    #     for i in range(1, readline.get_current_history_length() + 1):
-    #         print(f"{i}: {readline.get_history_item(i)}")
-
-    def load_aliases_file(self, filename):
-        """Loads alias definitions from a file (lines starting with .alias)."""
-        if not os.path.exists(filename):
-            self.logger.warning("Aliases file not found: %s", filename)
-            return
-        try:
-            with open(filename, "r") as f:
-                for line in f:
-                    line = line.strip()
-                    if line.startswith(".alias"):
-                        handle_alias_command(line, self.aliases)
-            self.logger.info("Loaded aliases file: %s", self.aliases_file)
-        except Exception as e:
-            print(f"Error loading aliases file {filename}: {e}")
-
-    def save_aliases_file(self, filename):
-        """Saves current aliases to a file using .alias syntax."""
-        try:
-            with open(filename, "w") as f:
-                for name, expr in sorted(self.aliases.items()):
-                    f.write(f".alias {name} = {expr}\n")
-            self.logger.info("Saved aliases to: %s", self.aliases_file)
-        except Exception as e:
-            print(f"Error saving aliases to {filename}: {e}")
-
-    def load_file(self, path, *, label=None, show_errors=True):
-        """Loads and executes each command from a file.
-
-        Args:
-            path (str): The file path.
-            label (str): Optional label for messages (e.g. "init file", ".reload", etc).
-            show_errors (bool): Whether to print errors from eval.
-        """
-        if not os.path.exists(path):
-            print(f"{label or 'File'} not found: {path}")
-            return
-
-        print(f"Loading {label or path}...")
-
-        try:
-            with open(path) as f:
-                for raw in f:
-                    line = raw.strip()
-                    if not line or line.startswith("#"):
-                        continue
-                    try:
-                        if line.startswith("."):
-                            if handle_alias_command(line, self.aliases):
-                                self.add_history_once(line)
-                                continue
-                            else:
-                                print(f"Unknown meta-command in file: {line}")
-                                continue
-                        try:
-                            expanded = expand_aliases(line, self.aliases)
-                        except ValueError as e:
-                            print(f"Alias error in file {path}: {e}")
-                            continue
-                        self.interpreter.eval(expanded)
-                        self.add_history_once(line)
-                    except Exception as e:
-                        if show_errors:
-                            print(f"Error in {label or path}: {e}")
-        except Exception as e:
-            print(f"Failed to read {label or path}: {e}")
 
     def process_line(self, line: str) -> bool:
         """Processes a single line of input. Returns True to continue, False to exit.
